+++ scripts/build.js
@@ -1,9 +1,12 @@
 const fs = require("fs-extra")
 const path = require("path")
 
-fs.writeFile(path.join(__dirname,'..','assets','build.json'),JSON.stringify({
+const BUILD_DIR = path.join(__dirname,'..','assets','build.json')
+const buildData = JSON.stringify({
 	date:new Date().toDateString()
-})).then(()=>{
+})
+
+fs.writeFile(BUILD_DIR,buildData).then(()=>{
 	console.log(`✨ Created build's JSON.`)
 }).catch((err)=>{
 	console.log(err)