+++ src/javascript/defaults/store/components/card.js
@@ -2,7 +2,6 @@ import { puffin } from '@mkenzo_8/puffin'
 import { Titles , Card} from '@mkenzo_8/puffin-drac'
 import Dialog from '../../../constructors/dialog'
 import Notification from '../../../constructors/notification'
-import requirePath from '../../../utils/require'
 import Endpoints from '../api/api.endpoints.js'
 import axios from 'axios'