+++ src/javascript/utils/directory.beautifier.js
@@ -1,5 +1,4 @@
-import requirePath from './require'
-const path = requirePath("path")
+const path = window.require("path")
 
 function beautifyDir(dir){
 	return path.normalize(dir).replace(/\\\\/g,'\\').replace(/"/g,'')