+++ src/javascript/utils/platform.js
@@ -1,8 +1,7 @@
-import requirePath from './require'
 import StaticConfig from 'StaticConfig'
 let AppPlatform = null;
 if( StaticConfig.data.appPlatform == 'auto' ){
-	AppPlatform = requirePath("electron").remote.process.platform
+	AppPlatform = window.require("electron").remote.process.platform
 }else{
 	AppPlatform = StaticConfig.data.appPlatform 
 }