+++ src/javascript/utils/require.js
@@ -1,5 +1,0 @@
-function requirePath(file){
-	return require(file)
-}
-
-export default requirePath